#18 Move sidenavbar into molecules folder
